Refactor test fixtures; update tests for new slug generation
