+++ capstone/capdb/tests/test_process_metadata.py
@@ -1,6 +1,9 @@
 import os
 import datetime
+import pytest
 
+from capdb.models import CaseMetadata, VolumeXML, CaseXML
+from capdb.tasks import create_case_metadata_from_all_vols
 from scripts import process_metadata
 from scripts.helpers import read_file
 
@@ -20,3 +23,20 @@ def test_get_case_metadata():
                     assert type(case_metadata["decision_date"]) is datetime.datetime
                     assert type(case_metadata["decision_date_original"]) is str
 
+@pytest.mark.django_db
+def test_create_case_metadata_from_all_vols(case_xml):
+    # get initial state
+    metadata_count = CaseMetadata.objects.count()
+    case_id = case_xml.metadata.case_id
+
+    # delete case metadata
+    case_xml.metadata.delete()
+    assert CaseMetadata.objects.count() == metadata_count - 1
+
+    # recreate case metadata
+    create_case_metadata_from_all_vols()
+
+    # check success
+    case_xml.refresh_from_db()
+    assert CaseMetadata.objects.count() == metadata_count
+    assert case_xml.metadata.case_id == case_id
