+++ capstone/capdb/tests/test_utils.py
@@ -1,24 +1,20 @@
 import pytest
-from capdb import utils
-from capdb.models import Reporter, Jurisdiction
+from django.utils.text import slugify
 
 
 @pytest.mark.django_db
-def test_generate_unique_slug(ingest_metadata):
-    unique_name = "Caselaw Access Project"
+def test_generate_unique_slug_manually(case):
+    old_slug = case.slug
+    case.pk = None
+    case.slug = None
+    case.save(slug_base=old_slug)
+    assert case.slug == old_slug + '-1'
 
-    slug1 = utils.generate_unique_slug(Reporter(), unique_name, field="full_name")
-    assert slug1 == "caselaw-access-project"
-
-    reporter = Reporter.objects.first()
-    reporter_name = reporter.full_name
-    slug2 = utils.generate_unique_slug(Reporter(), reporter_name, field="full_name")
-    assert slug2 == utils.slugify(reporter_name)
-
-    # method deals with collisions
-    jurisdiction = Jurisdiction(name='New. State.')
-    jurisdiction.slug = utils.generate_unique_slug(Jurisdiction(), jurisdiction.name)
-    jurisdiction.save()
-
-    new_unique_slug = utils.generate_unique_slug(Jurisdiction(), jurisdiction.name)
-    assert new_unique_slug != jurisdiction.slug
+@pytest.mark.django_db
+def test_generate_unique_slug_automatically(court):
+    old_slug = court.slug
+    assert old_slug == slugify(court.name_abbreviation)
+    court.pk = None
+    court.slug = None
+    court.save()
+    assert court.slug == old_slug + '-1'
